+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - Replaced "IBM Plex Sans" text font with "Inter" as there were reports for being difficult to read on dark background as it was too "thin" (specifically on Firefox).
     _If you are still having issues with the text readability, please open an issue with more details about your OS and browser to try to workaround it._
 
+- Serve fixed `Content-Type` for `.xlsx`, `.docx` and `.pptx` files to allow previews on iOS ([#7467](https://github.com/pocketbase/pocketbase/discussions/7467)).
+
 
 ## v0.37.5